cantidad del detalle de int a double

+++ lib/app/domain/models/erp/orden_compra/orden_compra_det.dart
@@ -19,7 +19,7 @@ class OrdenCompraDet {
   int codigoTipoCompra;
   String codArticulo;
   String nomArticulo;
-  int cantidad;
+  double cantidad;
   double precioUnitario;
   int codimpuesto;
   double valorImpuesto;
